+++ flexmeasures/api/dev/tests/conftest.py
@@ -1,6 +1,5 @@
 from datetime import timedelta
 
-from flask_security import SQLAlchemySessionUserDatastore
 import pytest
 
 from flexmeasures.data.models.generic_assets import GenericAssetType, GenericAsset
@@ -13,8 +12,7 @@ def setup_api_test_data(db, setup_roles_users):
     Set up data for API dev tests.
     """
     print("Setting up data for API v2.0 tests on %s" % db.engine)
-    add_gas_sensor(db, setup_roles_users["Test Supplier"])
-    give_prosumer_the_MDC_role(db)
+    add_gas_sensor(db, setup_roles_users["Test Prosumer User 2"])
 
 
 @pytest.fixture(scope="function")
@@ -25,8 +23,7 @@ def setup_api_fresh_test_data(fresh_db, setup_roles_users_fresh_db):
     print("Setting up fresh data for API dev tests on %s" % fresh_db.engine)
     for sensor in Sensor.query.all():
         fresh_db.delete(sensor)
-    add_gas_sensor(fresh_db, setup_roles_users_fresh_db["Test Supplier"])
-    give_prosumer_the_MDC_role(fresh_db)
+    add_gas_sensor(fresh_db, setup_roles_users_fresh_db["Test Prosumer User 2"])
 
 
 def add_gas_sensor(db, test_supplier):
@@ -38,6 +35,7 @@ def add_gas_sensor(db, test_supplier):
     incineration_asset = GenericAsset(
         name="incineration line",
         generic_asset_type=incineration_type,
+        account_id=test_supplier.account_id,
     )
     db.session.add(incineration_asset)
     db.session.flush()
@@ -49,13 +47,3 @@ def add_gas_sensor(db, test_supplier):
     )
     db.session.add(gas_sensor)
     gas_sensor.owner = test_supplier
-
-
-def give_prosumer_the_MDC_role(db):
-
-    from flexmeasures.data.models.user import User, Role
-
-    user_datastore = SQLAlchemySessionUserDatastore(db.session, User, Role)
-    test_prosumer = user_datastore.find_user(email="test_prosumer@seita.nl")
-    mdc_role = user_datastore.create_role(name="MDC", description="Meter Data Company")
-    user_datastore.add_role_to_user(test_prosumer, mdc_role)