+++ flexmeasures/api/v1_3/tests/test_api_v1_3.py
@@ -23,7 +23,7 @@ from flexmeasures.utils.calculations import integrate_time_series
 def test_get_device_message_wrong_event_id(client, message):
     asset = Asset.query.filter(Asset.name == "Test battery").one_or_none()
     message["event"] = message["event"] % (asset.owner_id, asset.id)
-    auth_token = get_auth_token(client, "test_prosumer@seita.nl", "testtest")
+    auth_token = get_auth_token(client, "test_prosumer_user@seita.nl", "testtest")
     get_device_message_response = client.get(
         url_for("flexmeasures_api_v1_3.get_device_message"),
         query_string=message,
@@ -54,7 +54,7 @@ def test_post_udi_event_and_get_device_message(
         asset_id = asset.id
         asset_owner_id = asset.owner_id
         message["event"] = message["event"] % (asset.owner_id, asset.id)
-        auth_token = get_auth_token(client, "test_prosumer@seita.nl", "testtest")
+        auth_token = get_auth_token(client, "test_prosumer_user@seita.nl", "testtest")
         post_udi_event_response = client.post(
             url_for("flexmeasures_api_v1_3.post_udi_event"),
             json=message,
@@ -126,7 +126,7 @@ def test_post_udi_event_and_get_device_message(
         asset_owner_id,
         asset_id,
     )
-    auth_token = get_auth_token(client, "test_prosumer@seita.nl", "testtest")
+    auth_token = get_auth_token(client, "test_prosumer_user@seita.nl", "testtest")
     get_device_message_response = client.get(
         url_for("flexmeasures_api_v1_3.get_device_message"),
         query_string=get_device_message,