+++ flexmeasures/auth/policy.py
@@ -1,0 +1,5 @@
+# TODO: use in access control lists, see https://github.com/SeitaBV/flexmeasures/issues/200
+PERMISSIONS = ["create", "read", "write", "delete"]
+
+ADMIN_ROLE = "admin"
+ADMIN_READER_ROLE = "admin-reader"