+++ flexmeasures/ui/utils/view_utils.py
@@ -11,6 +11,7 @@ from werkzeug.exceptions import BadRequest
 import iso8601
 
 from flexmeasures import __version__ as flexmeasures_version
+from flexmeasures.auth.policy import ADMIN_ROLE
 from flexmeasures.utils import time_utils
 from flexmeasures.ui import flexmeasures_ui
 from flexmeasures.data.models.user import User
@@ -31,7 +32,7 @@ def render_flexmeasures_template(html_filename: str, **variables):
     variables["show_queues"] = False
     if current_user.is_authenticated:
         if (
-            current_user.has_role("admin")
+            current_user.has_role(ADMIN_ROLE)
             or current_app.config.get("FLEXMEASURES_MODE", "") == "demo"
         ):
             variables["show_queues"] = True
@@ -79,7 +80,7 @@ def render_flexmeasures_template(html_filename: str, **variables):
     variables["user_is_logged_in"] = current_user.is_authenticated
     variables[
         "user_is_admin"
-    ] = current_user.is_authenticated and current_user.has_role("admin")
+    ] = current_user.is_authenticated and current_user.has_role(ADMIN_ROLE)
     variables[
         "user_is_anonymous"
     ] = current_user.is_authenticated and current_user.has_role("anonymous")